requirements.txt
aiogram==2.0.1
aiohttp==3.5.4
alembic==0.9.6
asn1crypto==0.24.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
blinker==1.4
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.6.1
dominate==2.3.1
elasticsearch==6.1.1
emoji==0.5.1
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
idna==2.8
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.10
jwt==0.6.1
ldap3==2.5.2
Mako==1.0.7
MarkupSafe==1.1.1
more-itertools==6.0.0
multidict==4.5.2
mysql==0.0.2
#mysql-connector==2.1.6
mysqlclient==1.4.2.post1
pluggy==0.9.0
py==1.8.0
pyaes==1.6.1
pyasn1==0.4.5
pycparser==2.19
PyJWT==1.5.3
pytest==4.3.0
python-dateutil==2.6.1
python-dotenv==0.10.1
python-editor==1.0.3
pytz==2018.9
redis==2.10.6
requests==2.21.0
rq==0.9.2
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.1
Telethon==1.6.2
typing-extensions==3.7.2
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
yarl==1.3.0
